AVS/requirements.txt
certifi==2021.10.8
easydict==1.10
einops==0.3.2
gpuinfo==1.0.0a7
imageio==2.22.4
ipdb==0.13.9
moviepy==1.0.3
numpy==1.20.3
opencv_python==4.5.3.56
pandas==1.3.3
Pillow==9.3.0
PyYAML==6.0
resampy==0.2.2
SoundFile==0.10.3.post1
timm==0.6.7
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tqdm==4.62.3
wandb==0.12.6
